disambiguate the pow call for gcc 3.1
svn path=/trunk/koffice/; revision=149645

+++ kspread/kspread_util.cc
@@ -275,7 +275,7 @@ int util_decodeColumnLabelText( const QString &_col )
     int counterColumn = 0;
     for ( uint i=0; i < _col.length(); i++ )
     {
-        counterColumn = (int) pow(26 , _col.length() - i - 1);
+        counterColumn = (int) std::pow(26.0 , static_cast<int>(_col.length() - i - 1));
         if( _col[i] >= 'A' && _col[i] <= 'Z' )
             col += counterColumn * ( _col[i].latin1() - 'A' + 1);  // okay here (Werner)
         else if( _col[i] >= 'a' && _col[i] <= 'z' )
